solution submitted by other developers
Big  Retailer  is looking for your help in maximizing sales and minimizing returns for their delectable inventory of baked goods. With over 1 million stores across Mexico, the transactional dataset for the Group Inventory Demand competition is huge. 

Accurately predicting each store's demand will make sure customers aren’t staring at empty shelves, while also reducing the amount spent on refunds to store owners with surplus product unfit for sale.

you will forecast the demand of a product for a given week, at a particular store. The dataset you are given consists of 9 weeks of sales transactions in Mexico. Every week, there are delivery trucks that deliver products to the vendors. Each transaction consists of sales and returns. Returns are the products that are unsold and expired. The demand for a product in a certain week is defined as the sales this week subtracted by the return next week.

The train and test dataset are split based on time.